+++ resources/lang/en/admin/resources.php
@@ -1,0 +1,20 @@
+<?php
+
+return [
+    'add' => [
+        'form' => 'Addition form',
+        'submit' => 'Add',
+        'success' => 'Resources added for the specified planet',
+        'title' => 'Add resources',
+    ],
+    'subtract' => [
+        'form' => 'Subtraction form',
+        'submit' => 'Subtract',
+        'success' => 'Resources subtracted for the specified planet',
+        'title' => 'Subtract resources',
+    ],
+    'deuterium' => 'Deuterium',
+    'crystal' => 'Crystal',
+    'metal' => 'Metal',
+    'planet_id' => 'Planet ID',
+];